Move LockableMixin to locking submodule
